% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_yaml_template.R
\name{clean_yaml_names}
\alias{clean_yaml_names}
\title{Cleans yaml names by removing redundant Folder and name portions}
\usage{
clean_yaml_names(yaml_names)
}
\arguments{
\item{yaml_names}{a vector of name entries from a flattened list}
}
\value{
Names cleaned to remove uninformative portions of flattened names
}
\description{
Cleans yaml names by removing redundant Folder and name portions
}
\examples{
clean_yaml_names(c("Folders.other.name", "Folders.logs.name"))
}
