% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_yaml_template.R
\name{collapsing_file_path}
\alias{collapsing_file_path}
\title{Turns a vector of path components into a complete file path}
\usage{
collapsing_file_path(vector_of_path_components)
}
\arguments{
\item{vector_of_path_components}{vector where each entry is part of path}
}
\value{
string of completed file path
}
\description{
Turns a vector of path components into a complete file path
}
\examples{
collapsing_file_path(c("a", "b", "c"))
}
