% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill_dataset_paths.R
\name{expand_dataset_path_to_file}
\alias{expand_dataset_path_to_file}
\title{Appends default file path to entry if its not already an absolute path}
\usage{
expand_dataset_path_to_file(current_file_entry, default_path)
}
\arguments{
\item{current_file_entry}{path to expected file location}

\item{default_to_consider}{the default path context for the current entry}
}
\value{
an absolute path to the file as a string
}
\description{
Appends default file path to entry if its not already an absolute path
}
\examples{
expand_dataset_path_to_file("C://scripts.sql", "")
expand_dataset_path_to_file("sql/script.sql", "S:/")
}
