% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill_dataset_paths.R
\name{extension_exists}
\alias{extension_exists}
\title{Check whether there is an extension in the file path}
\usage{
extension_exists(file_path)
}
\arguments{
\item{file_path}{string of a path for a file}
}
\value{
boolean on whether an extension was found
}
\description{
Check whether there is an extension in the file path
}
\examples{
extension_exists("test.xlsx")
}
