% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expand_default_paths.R
\name{extract_default_entries}
\alias{extract_default_entries}
\title{For a list, get entries holding default values}
\usage{
extract_default_entries(yaml_as_list)
}
\arguments{
\item{yaml_as_list}{a yaml document that has been loaded into a list}
}
\value{
list of entries whose name matches default format
}
\description{
For a list, get entries holding default values
}
\examples{
extract_default_entries(as.list(c("default_val" = "a", "c" = "d")))
}
