% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expand_default_paths.R
\name{extract_default_names}
\alias{extract_default_names}
\title{For a list, get the names matching default path}
\usage{
extract_default_names(yaml_as_list)
}
\arguments{
\item{yaml_as_list}{a yaml document that has been loaded into a list}
}
\value{
vector of names matching default format
}
\description{
For a list, get the names matching default path
}
\examples{
extract_default_names(as.list(c("default_val" = "a", "c" = "d")))
}
