% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_yaml_template.R
\name{extract_file_paths_from_names}
\alias{extract_file_paths_from_names}
\title{Turn's yaml name entries into matching file paths split on '.'}
\usage{
extract_file_paths_from_names(clean_folder_names, default_path)
}
\arguments{
\item{clean_folder_names}{a vector of names for the folder entries in list}

\item{default_path}{default path to add to relative folder path entry}
}
\value{
a vector of file paths for each folder
}
\description{
Turn's yaml name entries into matching file paths split on '.'
}
\examples{
extract_file_paths_from_names(c("other", "logs.test", "data.raw_data"), ".")
}
