% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_yaml_template.R
\name{get_default_entry_name}
\alias{get_default_entry_name}
\title{Adds default indicator to non default field name}
\usage{
get_default_entry_name(non_default_field_name)
}
\arguments{
\item{non_default_field_name}{non default field name}
}
\value{
string of matched default field name
}
\description{
Adds default indicator to non default field name
}
