% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill_dataset_paths.R
\name{get_extension}
\alias{get_extension}
\title{Gets file extension from a file path}
\usage{
get_extension(path_to_file)
}
\arguments{
\item{path_to_file}{string of a path for a file}
}
\value{
the extension ending of the file
}
\description{
Gets file extension from a file path
}
\examples{
get_extension("test.xlsx")
}
