% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_yaml_template.R
\name{get_non_default_entry_name}
\alias{get_non_default_entry_name}
\title{Removed default indicator from default field name}
\usage{
get_non_default_entry_name(default_field_name)
}
\arguments{
\item{default_field_name}{default version of a field name}
}
\value{
string of matched non-default field name
}
\description{
Removed default indicator from default field name
}
