% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expand_default_paths.R
\name{identify_default_entries}
\alias{identify_default_entries}
\title{Identifies entries in vector matching default variable format}
\usage{
identify_default_entries(entry_names)
}
\arguments{
\item{entry_names}{the names from entries in a yaml as a list}
}
\value{
Boolean vector indicating if entry matches default name format
}
\description{
Identifies entries in vector matching default variable format
}
\examples{
identify_default_entries(c("default_z", "a", "b", "c_default"))
}
