% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expand_default_paths.R
\name{identify_path_entries}
\alias{identify_path_entries}
\title{Identifies entries in vector match path variable format}
\usage{
identify_path_entries(entry_names)
}
\arguments{
\item{entry_names}{the names from entries in a yaml as a list}
}
\value{
Boolean vector indicating if entry matches path name format
}
\description{
Identifies entries in vector match path variable format
}
\examples{
identify_default_entries(c("z_path", "a", "b", "path_c"))
}
