% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_yaml_template.R
\name{prepare_dataset_yaml}
\alias{prepare_dataset_yaml}
\title{Title}
\usage{
prepare_dataset_yaml(yaml_data_as_list, default_path, dataset_list_updater)
}
\arguments{
\item{yaml_data_as_list}{yaml document loaded as a nested list}
}
\description{
Title
}
