% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_yaml_template.R
\name{prepare_folder_structure_yaml}
\alias{prepare_folder_structure_yaml}
\title{Processes a folder layout yaml to get paths to each directory}
\usage{
prepare_folder_structure_yaml(yaml_data_as_list, default_directory)
}
\arguments{
\item{yaml_data_as_list}{list containing data from loading yaml file}

\item{default_directory}{the directory to append before relative paths}
}
\value{
A list with entries being absolute paths to each folder in project
}
\description{
Processes a folder layout yaml to get paths to each directory
}
\examples{
prepare_folder_structure_yaml(list("Folders" =
list(list("data" = list("name" = "data_dir")))), ".")
}
