% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expand_default_paths.R
\name{process_default_path_entry}
\alias{process_default_path_entry}
\title{Fills empty path value with the default and expands relative paths}
\usage{
process_default_path_entry(path_to_check, path_default)
}
\arguments{
\item{path_to_check}{a path entry to expand to an absolute path}

\item{path_default}{path to use if missing an entry}
}
\value{
an absolute path to a file
}
\description{
Fills empty path value with the default and expands relative paths
}
\examples{
process_default_path_entry(NULL, ".")
process_default_path_entry("data", "C:/")
process_default_path_entry("data/test", "C:/")
}
