% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_yaml_template.R
\name{process_non_path_entry}
\alias{process_non_path_entry}
\title{Title}
\usage{
process_non_path_entry(entry_name, default_value, dataset_block)
}
\arguments{
\item{dataset_block}{}
}
\description{
Title
}
