% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill_dataset_paths.R
\name{replace_file_extension}
\alias{replace_file_extension}
\title{Replaces extisting file extension in a path with the provided one}
\usage{
replace_file_extension(file_path, new_extension)
}
\arguments{
\item{file_path}{string of a path for a file}

\item{new_extension}{extension to replace in existing file path}
}
\value{
a matching filepath but with the file extension swapped
}
\description{
Replaces extisting file extension in a path with the provided one
}
\examples{
replace_file_extension("test.xlsx", "parquet")
}
