% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_yaml_template.R
\name{update_data_path_with_name}
\alias{update_data_path_with_name}
\title{Title}
\usage{
update_data_path_with_name(
  path_to_data,
  default_data_path,
  file_name,
  data_file_type
)
}
\arguments{
\item{data_file_type}{}
}
\description{
Title
}
