% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_yaml_template.R
\name{update_disk_dataset_entry}
\alias{update_disk_dataset_entry}
\title{For a disk yaml data block, update all of the defaults}
\usage{
update_disk_dataset_entry(dataset_block, current_defaults)
}
\arguments{
\item{dataset_block}{a block of data for disk datasets}

\item{current_defaults}{a block of default options for entries}
}
\value{
A dataset block with disk-source specific defaults all applied
}
\description{
For a disk yaml data block, update all of the defaults
}
\examples{
update_disk_dataset_entry(list(), list())
}
