% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_yaml_template.R
\name{update_exec_dataset_list}
\alias{update_exec_dataset_list}
\title{Title}
\usage{
update_exec_dataset_list(yaml_data_as_list)
}
\arguments{
\item{yaml_data_as_list}{yaml document loaded as a nested list}
}
\description{
Title
}
