% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_yaml_template.R
\name{update_list_of_datasets}
\alias{update_list_of_datasets}
\title{Updates entries for every dataset in the yaml list}
\usage{
update_list_of_datasets(yaml_data_as_list, individual_dataset_processor)
}
\arguments{
\item{yaml_data_as_list}{yaml document loaded as a nested list}

\item{individual_dataset_processor}{a function to update a dataset block}
}
\value{
a yaml datalist with updated measures
}
\description{
Updates entries for every dataset in the yaml list
}
\examples{
simple_data_update <- function(dataset_block, dataset_defaults) {
  dataset_block$default <- "test_value"
  return(dataset_block)
}
test_yaml <- list(
  "default_script_path" = ".",
  "datasets_to_load" = list(
    list("name" = "dataset_one"),
    list("name" = "dataset_two")
  )
 )
update_list_of_datasets(test_yaml, simple_data_update)
}
