% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_yaml_template.R
\name{fill_list_default_paths}
\alias{fill_list_default_paths}
\title{Title}
\usage{
fill_list_default_paths(yaml_data_as_list, default_path_to_consider)
}
\arguments{
\item{yaml_data_as_list}{yaml document loaded as a nested list}

\item{default_path_to_consider}{}
}
\description{
Title
}
