% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_yaml_template.R
\name{prepare_dataset_yaml}
\alias{prepare_dataset_yaml}
\title{Updates default paths for the yaml data block}
\usage{
prepare_dataset_yaml(yaml_data_as_list, default_path, dataset_list_updater)
}
\arguments{
\item{yaml_data_as_list}{yaml document loaded as a nested list}

\item{default_path}{default path local to use to append to entries}

\item{dataset_list_updater}{function to update a list of datasets}
}
\value{
yaml data block with default paths filled in and appropriate names
}
\description{
Updates default paths for the yaml data block
}
\examples{
# Simple example with a small, generic, yaml dataset
default_path <- file.path("C:", "test")
simple_data_update <- function(x) {x$test <- "value"; return(x)}
test_yaml <- list(
  "default_script_path" = NULL,
  "datasets_to_load" = list(
    list("name" = "dataset_one"),
    list("name" = "dataset_two")
  )
 )
prepare_dataset_yaml(test_yaml, default_path, simple_data_update)
}
