% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_yaml_template.R
\name{update_data_path_with_name}
\alias{update_data_path_with_name}
\title{Uses standard names to update path to a data depending on default options}
\usage{
update_data_path_with_name(
  path_to_data,
  default_data_path,
  file_name,
  data_file_type
)
}
\arguments{
\item{path_to_data}{path to the expected data folder}

\item{default_data_path}{A default file path to use for dataset location}

\item{file_name}{the name of the file being considered}

\item{data_file_type}{the default file type to use if missing an extension}
}
\value{
an absolute path to for a data file
}
\description{
Uses standard names to update path to a data depending on default options
}
\examples{
update_data_path_with_name("raw_data", "./data", "test_data.sql", "parquet")
}
