% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_yaml_template.R
\name{update_non_path_entries}
\alias{update_non_path_entries}
\title{Title}
\usage{
update_non_path_entries(non_path_default_names, dataset_block)
}
\arguments{
\item{dataset_block}{}
}
\description{
Title
}
