% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_yaml_template.R
\name{update_online_list_of_datasets}
\alias{update_online_list_of_datasets}
\title{Go over the online datasets and update all dataset blocks}
\usage{
update_online_list_of_datasets(yaml_data_as_list)
}
\arguments{
\item{yaml_data_as_list}{yaml document loaded as a nested list}
}
\value{
An online yaml dataset as a list with all data set entries updated
}
\description{
Go over the online datasets and update all dataset blocks
}
\examples{
\dontrun{
update_online_list_of_datasets(list())
}
}
